requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
flask-login
flask-migrate
flask-cors
python-dotenv
pytesseract
pillow
flask-wtf
wtforms
email-validator
transformers
torch
pypdf
